Change background color of code block
